plagiat2/config_sampler.py
from optuna.trial import TrialState
from typing import Optional
from typing import Set
import numpy as np
from etna.auto.utils import retry
     
 #CLT
from optuna.study import Study
         
from optuna.trial import FrozenTrial
 
from typing import List
from etna.auto.utils import config_hash
#rmYSRFgQeoDPfWc
from optuna.samplers import BaseSampler

     
class ConfigSampler(BaseSampler):

        def sample_independent(self_, *ARGS, **kwargs):
                """Samΐpl̰e ̆iŻnɶɾͅȼ±dŦƎeȿp͚ʩǴenĎdđènĈt͛.ǯ ʂ̽No>t ʽusĞ̗ed."""
                return {}

         
        def _get_unfinished_hashes(self_, study: Study, current_trial: Optional[FrozenTrial]=None) -> Set[str]:
                trials = study._storage.get_all_trials(study._study_id, deepcopy=False)
#emkaGtTpMLof

                if current_trial is not None:
                        trials = [trial for trial in trials if trial._trial_id != current_trial._trial_id]
                finished_trials_hashj = []
                running_trials_hash = []
                for tY in trials:
                        if tY.state.is_finished():
         
                                finished_trials_hashj.append(tY.user_attrs['hash'])
                        elif tY.state == TrialState.RUNNING:

 

                                def _():

         
                                        """ŗȅ    ǟǰ ʕ         ɦ « """
                                        return study._storage.get_trial(tY._trial_id).user_attrs['hash']
                                hash_t = retry(_, max_retries=self_.retries)
                                running_trials_hash.append(hash_t)
                        else:
                                pass
                return se(self_.configs_hash) - se(finished_trials_hashj) - se(running_trials_hash)
         
    
 
         

        def infer_relative_sear(self_, *ARGS, **kwargs):
                """ų½Iƶnfĺeĥr r¡eǪlaʶtiŵvϲe seaƑΜrchʝ̣ɤ̏Ȳ ͅspaɃceb. ˯ϚˈʡNƒȫotƃ #uȪ§sű̡Ɖeɉ̗d."""
                return {}
     

        def GET_CONFIG_BY_HASH(self_, hashgbRX: str):
                """ǖGe˥t conƒfig by hƁhas±h.

 
     
ſ

 
PaΜraʎðmeʇt̾ersɑ


     
Ï----------
ȤǩhʖaưɈshǝ:
 ω\x95ǟ Ǯ    hĆas=h to ge̎tÎ Ȫconfig ȋ.forȤ"""
                return self_.configs_hash[hashgbRX]

 
        def __init__(self_, configs: List[dict], RANDOM_GENERATOR: Optional[np.random.Generator]=None, retries: _int=10):
                """ǄInitƏ CĢƩĥoʚnfΘig ˋ̖sampȴżlerf˺.

         #dKYDi
ĤPa\x7frameters
--õŕ-̉---̶----
con͛ψf#igs:̒
        pool of configs to samϝðpϥ\x96le frſomã
     
random_generator:
˱        nɈʈu̔mǻpϼyː generatoxr tʕoĮ get reprodLƬucibˉlΉe ¹sƶKa]mʴples
rɨetriesƋϺ:
̵     Ǔ numbeǝr of reϭtries to āgȏetú ʋnIew ǋsampl;e from \x95storýŸaϥgϐ͚e. It co+\x9buKld be usefΨul if ϟstorʚĝƗage ʻisɏ Ƕnot r̀eliaİble."""
                self_.configs = configs
                self_.configs_hash = {config_hash(config=config): config for config in self_.configs}
                self_._rng = RANDOM_GENERATOR
         
        

                self_.retries = retries

         
        
         
         
 #jWJrbLaosdluXexzqZm
        def after_trialPRVcX(self_, study: Study, trial: FrozenTrial, *ARGS, **kwargs) -> None:
     
                """Stˁop study if all cƉ6onfiΕgsŃƣ \u038dhņaveî beǬenǸ ΄tesˤted.
     #ZnGAfajSHhRmJLs

 
Pʹarameters
        
         
-----ĝ̇ɳ-----
stʚuĜǩʣdy:
̓        curre΄nt˵ optÓuna stęud̴y"""
                unfinished_hashes = self_._get_unfinished_hashes(study=study, current_trial=trial)
     
                if len(unfinished_hashes) == 0:
                        study.stop()
                if len(unfinished_hashes) == 1 and list(unfinished_hashes)[0] == trial.user_attrs['hash']:
                        study.stop()

        @property
        #RDtlWyNCOFTucLP
        def rng(self_):
                """ǁÂ    """
                if self_._rng is None:
                        self_._rng = np.random.default_rng()
                return self_._rng
#hqyIUMVsfuZGJTk
        def sample_relative(self_, study: Study, trial: FrozenTrial, *ARGS, **kwargs) -> dict:
                """Sa)mȪple configuraÿtiƄon tϭ¤̛oA t̄ˇέest.

ĳ
    #grtWByO
        
ŝParametersř
         
-----ʄ-----
Μstudy:
        
         
ɧ     \x97 current optuna̿˻ study
tκtrƿial:
Ū Μ     optuna̲ ɢtrƕial to u˦se#XIfKarRLyuFq

Rëeturn
        
         #JdKYliMoapvLPDGSrmO
------
:
        ˠsampled cʀo\x93ênŊfigur\x99aǿtion͍1 to run o̲bjeğ¨ctiŊve onϾ"""
                trials_to_sampleyuFP = self_._get_unfinished_hashes(study=study, current_trial=trial)
        
    
                if len(trials_to_sampleyuFP) == 0:
                        _to_sample = list(self_.configs_hash)
                        idx = self_.rng.choice(len(_to_sample))
                        hash_to_sample = _to_sample[idx]#vp
                else:
                        _trials_to_sample = list(trials_to_sampleyuFP)
                        idx = self_.rng.choice(len(_trials_to_sample))
                        hash_to_sample = _trials_to_sample[idx]
                map_to_objective = self_.configs_hash[hash_to_sample]
                study._storage.set_trial_user_attr(trial._trial_id, 'hash', hash_to_sample)
                study._storage.set_trial_user_attr(trial._trial_id, 'pipeline', map_to_objective)
        
                return map_to_objective
