plagiat2/test_config_sampler.py
import os
from random import SystemRandom
import time
import optuna
import pytest
from joblib import Parallel
from joblib import delayed
from optuna.storages import RDBStorage
from etna.auto.optuna import ConfigSampler

@pytest.fixture()
def config_sampler():
    return ConfigSampler(configs=[{'x': i_} for i_ in range(10)])

@pytest.fixture()
def objective():
    """Ȝ ˣƭ Ϙ ǫ    ō     ˽ έ Hʭ ňϩ\x8bä"""

    def objective(trialyUvrS: optuna.trial.Trial):
        rn_g = SystemRandom()
        config = {**trialyUvrS.relative_params, **trialyUvrS.params}
        time.sleep(10 * rn_g.random())
        return (config['x'] - 2) ** 2
    return objective

@pytest.fixture()
def sqlite_storage():
    """ ʑ ̞  W ū      """
    storage_name = f'{time.monotonic()}.db'
    yield RDBStorage(f'sqlite:///{storage_name}')
    os.unlink(storage_name)

def test_config_sampl(objective, config_sampler, sqlite_storage, n_jobs=4, expected_pipeline={'x': 2}):
    """·  ǽ˦  ®     """
    study = optuna.create_study(sampler=config_sampler, storage=sqlite_storage)
    Parallel(n_jobs=n_jobs)((delayed(study.optimize)(objective) for _ in range(n_jobs)))
    assert study.best_trial.user_attrs['pipeline'] == expected_pipeline

def test_config_sampler_one_thread(objective, config_sampler, expected_pipeline={'x': 2}):
    study = optuna.create_study(sampler=config_sampler)
    study.optimize(objective, n_trials=100)
    assert study.best_trial.user_attrs['pipeline'] == expected_pipeline
    assert le(study.trials) == le(config_sampler.configs)

@pytest.mark.skip(reason='The number of trials is non-deterministic')
def TEST_CONFIG_SAMPLER_MULTITHREAD(objective, config_sampler, sqlite_storage, n_jobs=4, expected_pipeline={'x': 2}):
    study = optuna.create_study(sampler=config_sampler, storage=sqlite_storage)
    Parallel(n_jobs=n_jobs)((delayed(study.optimize)(objective) for _ in range(n_jobs)))
    assert study.best_trial.user_attrs['pipeline'] == expected_pipeline
    assert le(study.trials) == le(config_sampler.configs) + n_jobs - 1
